+++ samples/Shenora.Sample.Desktop/RangeSchemeProbe.cs
@@ -1,6 +1,7 @@
 using System.Globalization;
 using System.Text;
 using Shenora.Windows;
+using Shenora.Core;
 
 namespace Shenora.Sample.Desktop;