+++ src/Shenora.Windows/WebView/WebViewHostOptions.cs
@@ -1,5 +1,6 @@
 using System.Drawing;
 using Microsoft.Web.WebView2.Core;
+using Shenora.Core;
 
 namespace Shenora.Windows;